// bio-energy-frontend/src/models/IBiomass.ts
export interface Biomass {
    id: string;
    cnCode: string | null;
    cpidCode: string | null;
    name: string | null;
    bulName: string | null;
    description: string | null;
  }
  export interface BiomassDTO {
    cnCode: string|null;
    cpidCode: string|null;
    name: string|null;
    bulName: string|null;
    description: string|null;
  }